# forgepay/api/openapi.yaml
openapi: "3.1.0"
info:
  title: ForgePay API
  version: "2026-04"
  description: |
    ForgePay unified payment API — accept card, stablecoin, and crypto payments;
    manage subscriptions; and query events. All amounts are integers in the
    smallest currency unit (cents for USD/EUR) unless noted.
  contact:
    name: ForgePay Developer Support
    url: https://myforgepay.com/docs
    email: api@myforgepay.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.myforgepay.com
    description: Production
  - url: https://api.sandbox.myforgepay.com
    description: Sandbox
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "Use your ForgePay API key: Authorization: Bearer sk_live_..."
  schemas:
...
    Subscription:
      type: object
      properties:
        id:           { type: string }
        customer_id:  { type: string }
        plan_id:      { type: string }
        status:       { type: string, enum: [active, cancelled, past_due, trialing] }
        created_at:   { type: string, format: date-time }
        current_period_start: { type: string, format: date-time }
        current_period_end:   { type: string, format: date-time }
      required: [id, customer_id, plan_id, status]
...
paths:
...
  /v1/subscriptions:
    post:
      summary: Create a subscription
      operationId: createSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customer_id: { type: string }
                plan_id:     { type: string }
                metadata:    { type: object }
              required: [customer_id, plan_id]
      responses:
        "200": { description: Subscription, content: { application/json: { schema: { $ref: "#/components/schemas/Subscription" } } } }
    get:
      summary: List subscriptions
      operationId: listSubscriptions
      parameters:
        - { name: customer_id, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Subscription" } }
  /v1/subscriptions/{subscription_id}:
    get:
      summary: Retrieve a subscription
      operationId: retrieveSubscription
      parameters:
        - { name: subscription_id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Subscription, content: { application/json: { schema: { $ref: "#/components/schemas/Subscription" } } } }
  /v1/subscriptions/{subscription_id}/cancel:
    post:
      summary: Cancel a subscription
      operationId: cancelSubscription
      parameters:
        - { name: subscription_id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Cancelled subscription, content: { application/json: { schema: { $ref: "#/components/schemas/Subscription" } } } }
...